feat: expose @zwave-js/config/safe
