+++ packages/config/src/Meters.ts
@@ -1,62 +1,9 @@
-import { isZWaveError, ZWaveError, ZWaveErrorCodes } from "@zwave-js/core";
-import { JSONObject, num2hex } from "@zwave-js/shared";
+import { JSONObject, num2hex } from "@zwave-js/shared/safe";
 import { entries } from "alcalzone-shared/objects";
 import { isObject } from "alcalzone-shared/typeguards";
-import { pathExists, readFile } from "fs-extra";
-import JSON5 from "json5";
-import path from "path";
-import {
-	configDir,
-	externalConfigDir,
-	hexKeyRegexNDigits,
-	throwInvalidConfig,
-} from "./utils";
+import { hexKeyRegexNDigits, throwInvalidConfig } from "./utils_safe";
 
 export type MeterMap = ReadonlyMap<number, Meter>;
-
-/** @internal */
-export async function loadMetersInternal(
-	externalConfig?: boolean,
-): Promise<MeterMap> {
-	const configPath = path.join(
-		(externalConfig && externalConfigDir()) || configDir,
-		"meters.json",
-	);
-
-	if (!(await pathExists(configPath))) {
-		throw new ZWaveError(
-			"The config file does not exist!",
-			ZWaveErrorCodes.Config_Invalid,
-		);
-	}
-
-	try {
-		const fileContents = await readFile(configPath, "utf8");
-		const definition = JSON5.parse(fileContents);
-		if (!isObject(definition)) {
-			throwInvalidConfig("meters", "the database is not an object");
-		}
-
-		const meters = new Map();
-		for (const [id, meterDefinition] of entries(definition)) {
-			if (!hexKeyRegexNDigits.test(id)) {
-				throwInvalidConfig(
-					"meters",
-					`found invalid key "${id}" at the root. Meters must have lowercase hexadecimal IDs.`,
-				);
-			}
-			const idNum = parseInt(id.slice(2), 16);
-			meters.set(idNum, new Meter(idNum, meterDefinition as JSONObject));
-		}
-		return meters;
-	} catch (e) {
-		if (isZWaveError(e)) {
-			throw e;
-		} else {
-			throwInvalidConfig("meters");
-		}
-	}
-}
 
 export class Meter {
 	public constructor(id: number, definition: JSONObject) {
